% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{add_empty_roi}
\alias{add_empty_roi}
\title{Represent missing ROIs in the dataset with a single row}
\usage{
add_empty_roi(experiment_folder, tracker_data, n = 20)
}
\description{
This is necessary so the results are always for 20 channels
regardless of whether the channel had an animal or not.
Doing so makes the results systematic and easily comparable across runs
}
