% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_polygon}
\alias{add_polygon}
\title{Mark when a piece of IDOC hardware was active
in the form of a polygon (usually a rectangle)}
\usage{
add_polygon(shape_data, color = "red", border = "black")
}
\description{
Mark when a piece of IDOC hardware was active
in the form of a polygon (usually a rectangle)
}
