% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_utils.R
\name{add_significance_marks}
\alias{add_significance_marks}
\title{Annotate significance of data}
\usage{
add_significance_marks(
  panel,
  test,
  annotation_df,
  y_annotation,
  vjust,
  textsize,
  map_signif_level,
  family,
  xmin = 1,
  xmax = 2,
  offset
)
}
\arguments{
\item{annotation_df}{(data.frame): contains columns stars and whatever
variable segregating datasets}

\item{y_annotation}{(numeric): Position along the y axis where to place
the annotation}

\item{offset}{(numeric): how many units to trim the line
connecting the compared datasets}

\item{panel:}{ggplot2 graph}

\item{test:}{function used to compute the p-value}
}
\description{
Annotate significance of data
}
