% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse_dataset}
\alias{analyse_dataset}
\title{Analyse an IDOC dataset}
\usage{
analyse_dataset(dataset, min_exits_required, min_time = 0, ...)
}
\arguments{
\item{min_exits_required}{Minimum number of exists required
when computing the preference index. Animals not reaching this number of exits
are not considered in the analysis}

\item{...}{Additional arguments to annotate_cross}
}
\description{
Take an IDOC dataset
}
\seealso{
annotate_cross

annotate_cross
}
