% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{annotate_cross}
\alias{annotate_cross}
\title{Annotate registered corsses with either appetitive or aversive behavior}
\usage{
annotate_cross(cross_data, event_data, treatment, type = "appetitive")
}
\arguments{
\item{cross_data}{dataframe where each row is an exit, defined by region_id, t and side}

\item{event_data}{dataframe where each row is an event, defined by stimulus, t_start t_end, side}
}
\value{

}
\description{
Annotate registered corsses with either appetitive or aversive behavior
}
