% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate_facet}
\alias{annotate_facet}
\title{Customize the facet label}
\usage{
annotate_facet(data, plot_preference_index = TRUE)
}
\arguments{
\item{data}{List with entries roi, controller, border, limits.
roi and controller must be data.tables generated by loading IDOC data.
Border is a TODO. Limits is a numeric vector with 2 positions, delineating
the boundaries of the x axis}

\item{plot_preference_index}{Whether to show the scored preference index
with the region id on the facet label (TRUE), or just the region id (FALSE)}
}
\description{
Customize the facet label
}
