% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{base_plot}
\alias{base_plot}
\title{Generate base IDOC plot upon which
more annotation layers can be added}
\usage{
base_plot(data, limits)
}
\arguments{
\item{data}{List with entries roi, controller, border, limits.
roi and controller must be data.tables generated by loading IDOC data.
Border is a TODO. Limits is a numeric vector with 2 positions, delineating
the boundaries of the x axis}

\item{limits}{Position of ticks in space axis}
}
\value{
gg ggplot2 object
}
\description{
The base plot contains all individual panels,
dots to mark the fly positions
}
