% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior_analyzer.R
\name{BehaviorAnalyzer}
\alias{BehaviorAnalyzer}
\title{Process the results of an experiment performed using the FSL learning memory setup}
\description{
Process the results of an experiment performed using the FSL learning memory setup

Process the results of an experiment performed using the FSL learning memory setup
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BehaviorAnalyzer$new()}}
\item \href{#method-read_experiment}{\code{BehaviorAnalyzer$read_experiment()}}
\item \href{#method-px2mm}{\code{BehaviorAnalyzer$px2mm()}}
\item \href{#method-program_name}{\code{BehaviorAnalyzer$program_name()}}
\item \href{#method-borders}{\code{BehaviorAnalyzer$borders()}}
\item \href{#method-add_period_column}{\code{BehaviorAnalyzer$add_period_column()}}
\item \href{#method-make_side_independent}{\code{BehaviorAnalyzer$make_side_independent()}}
\item \href{#method-compute_side}{\code{BehaviorAnalyzer$compute_side()}}
\item \href{#method-compute_index}{\code{BehaviorAnalyzer$compute_index()}}
\item \href{#method-find_period_edge}{\code{BehaviorAnalyzer$find_period_edge()}}
\item \href{#method-find_relevant_periods}{\code{BehaviorAnalyzer$find_relevant_periods()}}
\item \href{#method-prepare_rect_data}{\code{BehaviorAnalyzer$prepare_rect_data()}}
\item \href{#method-add_rectangles}{\code{BehaviorAnalyzer$add_rectangles()}}
\item \href{#method-prepare_index_data}{\code{BehaviorAnalyzer$prepare_index_data()}}
\item \href{#method-plot_trace_with_pin_events}{\code{BehaviorAnalyzer$plot_trace_with_pin_events()}}
\item \href{#method-events_over_time_plot}{\code{BehaviorAnalyzer$events_over_time_plot()}}
\item \href{#method-run}{\code{BehaviorAnalyzer$run()}}
\item \href{#method-clone}{\code{BehaviorAnalyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$new(
  experiment_folder,
  index_function = LeMDTr::preference_index,
  decision_zone_mm = 10,
  min_exits_required = 5,
  max_time_minutes = Inf,
  annotation = "",
  A = NULL,
  B = NULL,
  selected_flies = 1:20,
  debug = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiment_folder}}{character, path to results folder of an experiment}

\item{\code{index_function}}{function, returns an satistic based on a character vector of states}

\item{\code{decision_zone_mm}}{numeric, widht of the decision in mm}

\item{\code{min_exits_required}}{integer, minimum number of crossing events required to return an index}

\item{\code{max_time_minutes}}{numeric, drop all data after this amount of time since the start of the recording}

\item{\code{annotation}}{character, text to write in the top left corner of the plot}

\item{\code{A}}{character, name of odour in valves connected to bottle A}

\item{\code{B}}{character, name of odour in valves connected to bottle B}

\item{\code{selected_flies}}{integer vector, only these \code{region_id}s will be included in the plot
Add to lemdt_result a column called period reflecting the state of the odour valves}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_experiment"></a>}}
\subsection{Method \code{read_experiment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$read_experiment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-px2mm"></a>}}
\subsection{Method \code{px2mm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$px2mm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-program_name"></a>}}
\subsection{Method \code{program_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$program_name(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-borders"></a>}}
\subsection{Method \code{borders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$borders(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_period_column"></a>}}
\subsection{Method \code{add_period_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$add_period_column()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{periods_dt}}{A data.table with a t column and any amount of binary columns capturing the state of the setup}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table with a new column called period summarising the state of all the binary columns in a single character
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_side_independent"></a>}}
\subsection{Method \code{make_side_independent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$make_side_independent()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_side"></a>}}
\subsection{Method \code{compute_side()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$compute_side()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_index"></a>}}
\subsection{Method \code{compute_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$compute_index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_period_edge"></a>}}
\subsection{Method \code{find_period_edge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$find_period_edge(
  dt,
  edge = "start",
  column_name = "period_id"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_relevant_periods"></a>}}
\subsection{Method \code{find_relevant_periods()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$find_relevant_periods(
  column_name = "period_id",
  min_datapoints = 100
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_rect_data"></a>}}
\subsection{Method \code{prepare_rect_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$prepare_rect_data(elshock_periods = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lemdt_result}}{A data.table with columns t, period_id}

\item{\code{elshock}}{Nothing for now}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table with columns xmin, xmax, ymin, ymax, fill and annotation, fit for geom_rect() and geom_label()
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_rectangles"></a>}}
\subsection{Method \code{add_rectangles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$add_rectangles(p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_index_data"></a>}}
\subsection{Method \code{prepare_index_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$prepare_index_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_trace_with_pin_events"></a>}}
\subsection{Method \code{plot_trace_with_pin_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$plot_trace_with_pin_events(annot_on_side = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-events_over_time_plot"></a>}}
\subsection{Method \code{events_over_time_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$events_over_time_plot(
  experiment_folder,
  period_column = "period"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{periods_dt}}{A data.table with two columns: t and period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BehaviorAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
