% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{build_filename}
\alias{build_filename}
\title{Module to infer names of .csv files based on metadata+key or experiment_folder+key
Build a filename based on the metadata and a key}
\usage{
build_filename(
  result_folder,
  metadata,
  key = NULL,
  suffix = "",
  extension = ".csv"
)
}
\arguments{
\item{result_folder}{Path to a folder where to save results. If not passaed, it is set to experiment_folder}

\item{metadata}{data.table with all information about the experiment
i.e. when it was done, etc}

\item{key}{Name of a table in the resulting database}

\item{suffix}{A character (word) that will differentiate resulting output
from other output produced on the same experiment folder.
Useful for segmenting the experiment into different periods of interest
and analyzing these periods separately.}

\item{extension}{Extension of file (.csv, ...)}
}
\value{
Filename with systematic metainformation and key to differentiate
it from other files from the same experiment
}
\description{
Module to infer names of .csv files based on metadata+key or experiment_folder+key
Build a filename based on the metadata and a key
}
