% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{center_around_median}
\alias{center_around_median}
\title{Set the median x position to 0}
\usage{
center_around_median(x)
}
\arguments{
\item{x}{Vector of animal positions}
}
\description{
Set the median x position to 0
}
