% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{center_dataset}
\alias{center_dataset}
\title{Center the trace of movement along x axis so x=0 is set
at the center of the chamber
Estimating the center via Computer Vision is very costly because the center
does not look similar across chambers (some chambers are wider than others, etc)
For that reason, if you as user wish maximum accuracy,
you should run the midline-detector/main.py script, available here
https://github.com/shaliulab/midline-detector
like so:
python main <--experiment-folder EXPERIMENT_FOLDER>}
\usage{
center_dataset(experiment_folder, tracker_data, infer = FALSE)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{tracker_data}{IDOC tracker dataset where every row captures the position of one animal at one timepoint}

\item{infer}{If false, read coords of roi centers from ROI_CENTER file
otherwise estimate based on fly behavior}
}
\description{
Center the trace of movement along x axis so x=0 is set
at the center of the chamber
Estimating the center via Computer Vision is very costly because the center
does not look similar across chambers (some chambers are wider than others, etc)
For that reason, if you as user wish maximum accuracy,
you should run the midline-detector/main.py script, available here
https://github.com/shaliulab/midline-detector
like so:
python main <--experiment-folder EXPERIMENT_FOLDER>
}
