% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{check_api_version}
\alias{check_api_version}
\title{Check whether the user is running deprecated code}
\usage{
check_api_version(treatments)
}
\arguments{
\item{treatments}{Vector of treatments.
If named and names contain the word treatment,
it is likely that the user is running old API code}
}
\description{
Check whether the user is running deprecated code
}
