% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{clean_empty_roi}
\alias{clean_empty_roi}
\title{Undo the effect of add_empty_roi}
\usage{
clean_empty_roi(tracker_data, minimum = 100)
}
\description{
Useful to remove dummy rois or rois with noise
Works only if not testing. During testing
we may want to pass small (noise-like) datasets
}
