% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference.R
\name{compute_preference_data}
\alias{compute_preference_data}
\title{Wrapper around overlap_cross_events for both apetitive and aversive behavior}
\usage{
compute_preference_data(cross_data, rectangle_data, CSplus, CSminus)
}
\description{
Wrapper around overlap_cross_events for both apetitive and aversive behavior
}
