% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{construct_animal_id}
\alias{construct_animal_id}
\title{Give each animal a unique id based on the run id of the experiment/machine
and its position on the machine#'}
\usage{
construct_animal_id(experiment_folder, region_id)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{region_id}{Position of the animal in the machine}
}
\description{
Give each animal a unique id based on the run id of the experiment/machine
and its position on the machine#'
}
