% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference-index.R
\name{count_decision_zone_is_skipped}
\alias{count_decision_zone_is_skipped}
\title{Count how many times D is skipped on either direction}
\usage{
count_decision_zone_is_skipped(pos)
}
\arguments{
\item{pos}{}
}
\value{
Integer with RL and LR counts
}
\description{
Count how many times D is skipped on either direction
}
