% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference-index.R
\name{count_exits}
\alias{count_exits}
\title{Count exits from the decision zone to either side of a chamber}
\usage{
count_exits(pos)
}
\arguments{
\item{pos}{A character vector encoding the position of the fly in the chamber (L/D/R)}
}
\value{
A list wit the number of exits from D->R (right exit), left and total (left+right)
}
\description{
Count exits from the decision zone to either side of a chamber
}
\examples{
pos <- c('L','L','L','D','D','D','L','L',"L",'D','D','R','R','R','D')
count_exits(pos)

}
