% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_detector}
\alias{cross_detector}
\title{Detect crosses of the decision zone}
\usage{
cross_detector(tracker_data, border, side = c(-1, 1))
}
\arguments{
\item{tracker_data}{IDOC tracker dataset where every row captures the position of one animal at one timepoint}

\item{border}{Number of pixels from center to decision zone edge}

\item{side}{Either LEFT (-1) or RIGHT (1)}
}
\description{
Report timepoints when the animal crossed on either direction
(in/out) the decision zone on the passed \code{side}.
The decision zone is \code{border} pixels far from the center
}
