% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{define_rectangle_all}
\alias{define_rectangle_all}
\title{Wrapper around define_rectangle for several stimulus}
\usage{
define_rectangle_all(dataset)
}
\arguments{
\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}
}
\description{
Wrapper around define_rectangle for several stimulus
}
