% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{define_rectangles}
\alias{define_rectangles}
\title{Wrapper around define_rectangle for several hardware}
\usage{
define_rectangles(dataset)
}
\arguments{
\item{dataset}{A wide data.frame of hardware and a t column}
}
\description{
Wrapper around define_rectangle for several hardware
}
