% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-unique-periods.R
\name{define_unique_periods}
\alias{define_unique_periods}
\title{TODO Rewrite this crap}
\usage{
define_unique_periods(periods_dt)
}
\arguments{
\item{periods_dt}{A data.table with columns t (time) and period (character) which captures the state of the setup}
}
\description{
TODO Rewrite this crap
}
