% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{document_plot}
\alias{document_plot}
\title{Annotate experiment metadata on plot for documentation
Users can write down particular features of their experiment
in the subtitle (frequency, duty cycle, odours, paradigms, etc)
A default title with the run id is generated if the experiment_folder is passed
A default caption with the current timestamp is placed on the bottom right}
\usage{
document_plot(gg, experiment_folder = NULL, ...)
}
\arguments{
\item{gg}{ggplot2 object}

\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{...}{Extra arguments to ggplot2::labs}
}
\value{
document_gg("return")
}
\description{
Annotate experiment metadata on plot for documentation
Users can write down particular features of their experiment
in the subtitle (frequency, duty cycle, odours, paradigms, etc)
A default title with the run id is generated if the experiment_folder is passed
A default caption with the current timestamp is placed on the bottom right
}
\seealso{
\code{\link[ggplot2:labs]{ggplot2::labs()}}
}
