% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{document_script}
\alias{document_script}
\title{Make a backup of a file}
\usage{
document_script(src_file = NULL, dst_folder)
}
\arguments{
\item{src_file}{Path of file to be copied.
If NULL, take current file (works on RStudio)}

\item{dst_folder}{Path of destination folder}
}
\description{
This is useful for keeping a documentation of the R code
used to produce some results based on some data
}
