% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_dataset}
\alias{export_dataset}
\title{Export analysis results to experiment_folder}
\usage{
export_dataset(experiment_folder, dataset, analysis, ...)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}

\item{analysis}{List of processed data including:
\itemize{
\item annotation: every row is a decision zone exit and its appetitive / aversive
value
\item pi: preference index scored by each animal
}}

\item{...}{Extra arguments to export_summary}

\item{result_folder}{Path to a folder where to save results. If not passaed, it is set to experiment_folder}

\item{suffix}{A character (word) that will differentiate resulting output
from other output produced on the same experiment folder.
Useful for segmenting the experiment into different periods of interest
and analyzing these periods separately.}

\item{summary}{Whether to export a summary or not}
}
\description{
Generate the preference index table
Generate the summary table
}
\seealso{
\code{\link[=export_pi_summary]{export_pi_summary()}}

\code{\link[=export_summary]{export_summary()}}
}
