% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_pi_summary}
\alias{export_pi_summary}
\title{Export the preference index to a csv file}
\usage{
export_pi_summary(experiment_folder, pi, output_csv = NULL)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{pi}{Dataframe to be exported.
Should contain a column named preference_index}

\item{output_csv}{Path of csv output. If not provided,
a new file in the experiment folder with key SUMMARY is produced}
}
\description{
Export the preference index to a csv file
}
