% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_summary}
\alias{export_summary}
\title{Export a single csv with all key data for analysis and plotting outside of R}
\usage{
export_summary(experiment_folder)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}
}
\description{
Export a single csv with all key data for analysis and plotting outside of R
}
