% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference-index.R
\name{find_decision_zone_non_traversed}
\alias{find_decision_zone_non_traversed}
\title{Find t when D is skipped}
\usage{
find_decision_zone_non_traversed(t, pos)
}
\arguments{
\item{t}{}

\item{pos}{}
}
\value{
Numeric with t for each transition without D
}
\description{
Find t when D is skipped
}
