% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{find_exits}
\alias{find_exits}
\title{TODO Write documentation!}
\usage{
find_exits(
  tracker_data,
  border,
  side = c(-1, 1),
  cross_detector_FUN = cross_detector,
  mask_FUN = seconds_mask,
  ...
)
}
\arguments{
\item{border}{Distance in pixels from center to decision zone edge}

\item{side}{Integer either -1 (left) or 1 (right)}

\item{cross_detector_FUN}{Cross detector function}

\item{mask_FUN}{Masking function to ignore certain crosses given some heuristic}
}
\description{
TODO Write documentation!
}
\seealso{
seconds_mask

cross_detector
}
