% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{find_file}
\alias{find_file}
\title{Find a file with matching key in a experiment_folder}
\usage{
find_file(experiment_folder, key)
}
\arguments{
\item{experiment_folder}{Path to a folder with IDOC results}

\item{key}{String matching a file key (VAR_MAP, METADATA, ROI_1, ...)}
}
\value{
character
}
\description{
Find a file with matching key in a experiment_folder
}
