% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{find_rois}
\alias{find_rois}
\title{Return list of ROI .csv files in the passed experiment folder
Paths are absolute}
\usage{
find_rois(experiment_folder)
}
\description{
Return list of ROI .csv files in the passed experiment folder
Paths are absolute
}
