% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{format_text}
\alias{format_text}
\title{Format text so numbers on it are rendered consistently}
\usage{
format_text(text)
}
\arguments{
\item{text}{Character text to be formatted}
}
\value{
Formatted character
}
\description{
Format text so numbers on it are rendered consistently
}
