% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{generate_toy_dataset}
\alias{generate_toy_dataset}
\title{Generate a toy dataset}
\usage{
generate_toy_dataset(result_folder = NULL, paradigm = NULL, ...)
}
\arguments{
\item{result_folder}{Path to the destination folder where toy dataset should be stored
It is created if not available}

\item{paradigm}{data.table containing a paradigm i.e. a list of controller events}

\item{...}{Additional arguments for toy_roi_all}
}
\description{
Generate a toy dataset
}
\seealso{
toy_roi_all

toy_controller
}
