% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{get_event_data}
\alias{get_event_data}
\title{Obtain a tidy data frame of events from a loaded dataset}
\usage{
get_event_data(dataset)
}
\arguments{
\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}
}
\value{
data.frame where every row registers the activation of an IDOC stimulus
}
\description{
Obtain a tidy data frame of events from a loaded dataset
}
