% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{get_roi_center}
\alias{get_roi_center}
\title{Read the x coordinate of the center of the rois
This is useful for precise delineation of the decision zone
The function expects the ROI_CENTER and ROI_MAP files to exist}
\usage{
get_roi_center(experiment_folder)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}
}
\description{
Read the x coordinate of the center of the rois
This is useful for precise delineation of the decision zone
The function expects the ROI_CENTER and ROI_MAP files to exist
}
