% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masks.R
\docType{data}
\name{global_mask}
\alias{global_mask}
\title{Masks representing onset of condtioned stimuli}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
global_mask
}
\description{
Each vector represents a time interval (begin, end)
which can be used to exclude any stimuli outside of it
when counting decision zone exits
}
\keyword{datasets}
