% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{infer_decision_zone_exits}
\alias{infer_decision_zone_exits}
\title{Wrapper around gather_cross_data for left and right}
\usage{
infer_decision_zone_exits(
  roi_data,
  border = border,
  cross_detector_FUN = cross_detector
)
}
\description{
Wrapper around gather_cross_data for left and right
}
