% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{interpolate_controller}
\alias{interpolate_controller}
\title{Interpolate controller recordings
Interpolate the recordings registered in the controller data
by adding new interpolated ones, one for each new timepoint in index
The interpolation is done by carrying forward the last observation}
\usage{
interpolate_controller(controller_data, index)
}
\arguments{
\item{controller_data}{data.frame with recordings of IDOC status at several timepoints,
one timepoint per row. Wide format}

\item{index}{data.frame with a column named t containing timepoints
not available in controller_data (but contained within its min-max interval)}
}
\description{
Interpolate controller recordings
Interpolate the recordings registered in the controller data
by adding new interpolated ones, one for each new timepoint in index
The interpolation is done by carrying forward the last observation
}
