% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot}
\alias{iplot}
\title{Generate base IDOC plot upon which
more annotation layers can be added}
\usage{
iplot(experiment_folder, data, limits, run_id = NULL)
}
\description{
Generate base IDOC plot upon which
more annotation layers can be added
}
