% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_idoc_metadata}
\alias{link_idoc_metadata}
\title{Find the raw data for an experiment in an IDOC database
The found folder is saved as a new column of the metadata called idoc_folder
metadata must be a data frame with columns:
\itemize{
\item Files: the basename of the IDOC folder (without the parent folder)
}}
\usage{
link_idoc_metadata(metadata, result_dir, verbose = TRUE)
}
\description{
result_dir must be a directory with the following structure:
One folder for every month whene at least one experiment was done. The folder is called YYYY-MM-idoc_data
Inside the month folders, there must be one subfolder for every experiment group
Example
result_dir
|
|--- 2022-01-idoc_data
|--- 2022-01-10_13-00-00 (FOO_STM)
|--- 2022-01-10_13-00-00
|--- 2022-01-10_13-10-00
|--- 2022-01-10_13-20-00
|--- 2022-01-10_14-00-00 (BAR_STM)
|--- 2022-01-10_14-00-00
|--- 2022-01-10_14-10-00
|--- 2022-01-10_14-20-00
|--- 2022-02-idoc_data
|--- 2022-02-12_13-00-00 (FOO)
|---
|---
|---
|--- 2022-02-13_14-00-00 (BAR)
|---
|---
|---
}
