% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_metadata}
\alias{load_metadata}
\title{Load the METADATA table into R}
\usage{
load_metadata(experiment_folder)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}
}
\description{
Load the METADATA table into R
}
