% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{load_rois}
\alias{load_rois}
\title{Load data in ROI csv tables into a single R tibble}
\usage{
load_rois(experiment_folder)
}
\arguments{
\item{experiment_folder}{Path to a folder with IDOC results}
}
\value{
tibble
}
\description{
Data saved by IDOC to .csv files needs to be read into R
}
