% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{load_systematic_rois}
\alias{load_systematic_rois}
\title{Load data from n channels
and expose missing data if needed}
\usage{
load_systematic_rois(..., n = 20)
}
\arguments{
\item{...}{Arguments to load_rois}

\item{n}{Number of channels desired}
}
\value{
tibble
}
\description{
Load data from n channels
and expose missing data if needed
}
