% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_varmap}
\alias{load_varmap}
\title{Load the VAR_MAP table into R}
\usage{
load_varmap(experiment_folder)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}
}
\description{
Load the VAR_MAP table into R
}
