% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{make_outputs}
\alias{make_outputs}
\title{Saves data (.csv) and plots for characterization of sleep properties}
\usage{
make_outputs(dt, variable, folder, prefix = "", intervals = default_intervals)
}
\arguments{
\item{dt}{Ethoscope data in 10 second bin format}

\item{variable}{One of "asleep" or "interactions", it determines which animal property will be characterized}

\item{folder}{Where to save the results}

\item{prefix}{Optional, the output files will have this string prefixed to their names}

\item{intervals}{Which zt intervals to analyse to produce traces and summary statistics}
}
\description{
Saves data (.csv) and plots for characterization of sleep properties
}
