% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_rectangle}
\alias{make_rectangle}
\title{Generate a rectangle object
that will mark a piece of IDOC hardware was active}
\usage{
make_rectangle(shape_data, color = "red", border = "black", alpha = 0.4)
}
\arguments{
\item{shape_data}{Data table where every row is a corner of a polygon
The group field is the same for all corners of the same polygon
x and y show where on the plot the corner is
The color and transparency is provided by the other function arguments.}

\item{color}{Rectangle fill color}

\item{border}{Rectangle border color}

\item{alpha}{Rectangle transparency}
}
\value{
ggplot2 geom_polygon
}
\description{
Generate a rectangle object
that will mark a piece of IDOC hardware was active
}
