% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{make_summary}
\alias{make_summary}
\title{Put the most important data into a wide format table}
\usage{
make_summary(tracker_data, controller_data)
}
\arguments{
\item{tracker_data}{IDOC tracker dataset where every row captures the position of one animal at one timepoint}

\item{controller_data}{data.frame with recordings of IDOC status at several timepoints,
one timepoint per row. Wide format}
}
\value{
data.table
}
\description{
Put the most important data into a wide format table
}
