% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_analysis_mask}
\alias{mark_analysis_mask}
\title{Mark the analysis mask}
\usage{
mark_analysis_mask(gg, analysis_mask)
}
\arguments{
\item{gg}{ggplot2 object}

\item{analysis_mask}{Numeric vector of length 2
containing start and end timepoint of mask
in seconds since the start of the experiment
Every behavior outside of the mask is ignored}
}
\description{
Mark the time interval of the experiment for which
the preference computation is performed
This is shown with a yellow rectangle in the plot
}
\seealso{
seconds_mask

cross_detector

mark_analysis_mask
}
