% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_crosses}
\alias{mark_crosses}
\title{Mark decision zone exit events (crosses)}
\usage{
mark_crosses(gg, crossing_data)
}
\arguments{
\item{gg}{ggplot2 object}
}
\description{
Mark decision zone exit events (crosses)
}
