% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_decision_zone}
\alias{mark_decision_zone}
\title{Draw a dashed line to signal the decision zone}
\usage{
mark_decision_zone(gg, border, center_alpha = 0.2, orientation = "y")
}
\arguments{
\item{gg}{ggplot2 object}

\item{border}{Pixels between center and decision zone}

\item{center_alpha}{Transparency of line marking the center of the chamber}
}
\value{
gg ggplot2 object
}
\description{
Draw a dashed line to signal the decision zone
}
