% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_space}
\alias{mark_space}
\title{Mark space
Place the ticks on the space axis so visualization is convenient}
\usage{
mark_space(limits, gg, extra = c(0), orientation = "y")
}
\arguments{
\item{limits}{Position of ticks in space axis}

\item{gg}{ggplot2 object}

\item{extra}{Position of ticks in space besides limits}
}
\value{
gg ggplot2 object
}
\description{
Mark space
Place the ticks on the space axis so visualization is convenient
}
