% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_stimuli}
\alias{mark_stimuli}
\title{Mark stimuli / treatment action over time with rectangles on the plot}
\usage{
mark_stimuli(gg, analysis, colors)
}
\arguments{
\item{gg}{ggplot2 object}

\item{analysis}{List of processed data including:
\itemize{
\item annotation: every row is a decision zone exit and its appetitive / aversive
value
\item pi: preference index scored by each animal
}}

\item{colors}{Named character vector where values are colors and names
are treatments. It establishes the color used to represent the treatments
on the plot}
}
\value{
ggplot2 object
}
\description{
Mark stimuli / treatment action over time with rectangles on the plot
}
