% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_time}
\alias{mark_time}
\title{Mark time with a custom frequency}
\usage{
mark_time(data, gg, freq = 60, downward = TRUE)
}
\arguments{
\item{data}{List with entries roi, controller, border, limits.
roi and controller must be data.tables generated by loading IDOC data.
Border is a TODO. Limits is a numeric vector with 2 positions, delineating
the boundaries of the x axis}

\item{gg}{ggplot2 object}

\item{freq}{Number of seconds in between axis ticks, 60 by default}

\item{downward}{Whether the 0 should be at the top (TRUE)
or the bottom (FALSE) of the plot}
}
\value{
ggplot2 object
}
\description{
Place the ticks on the time axis with an intuitive frequency
}
