% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_cross_events.R
\name{overlap_cross_events}
\alias{overlap_cross_events}
\title{@export}
\usage{
overlap_cross_events(
  cross_data,
  event_data,
  type = "preference",
  mask_FUN = NULL,
  mirror = FALSE,
  ...
)
}
\description{
@export
}
