% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{pad_t0}
\alias{pad_t0}
\title{Make sure the timeseries starts at 0 and
all dependent variables are also set to 0}
\usage{
pad_t0(data)
}
\arguments{
\item{data}{A timeseries data.table with a column named t
representing time in a numeric format and assuming t starts at exactly 0}
}
\description{
Make sure the timeseries starts at 0 and
all dependent variables are also set to 0
}
