% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{parse_side}
\alias{parse_side}
\title{Parse side from the hardware name}
\usage{
parse_side(stimulus)
}
\arguments{
\item{stimulus}{Name of stimulus to be parsed.
Must contain LEFT or RIGHT on its name}
}
\description{
Return -1 for left and 1 for right
}
