% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idocr.R
\name{pipeline}
\alias{pipeline}
\title{Analyse plot and export dataset}
\usage{
pipeline(
  experiment_folder,
  dataset,
  min_exits_required,
  mask_duration,
  analysis_mask = NULL,
  ...
)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}

\item{min_exits_required}{Minimal number of exits
to be considered for significance}

\item{mask_duration}{Exits happening this amount of seconds after the last one are ignored}

\item{analysis_mask}{Numeric vector of length 2
containing start and end timepoint of mask
in seconds since the start of the experiment
Every behavior outside of the mask is ignored}

\item{...}{Extra arguments to plot_dataset}
}
\value{
ggplot2 object
}
\description{
Take an IDOC dataset
}
\details{
A corssing dataset where every record represents a decision zone
exit of one animal is required to show the exits in the plot
}
\seealso{
\code{\link[=annotate_cross]{annotate_cross()}}
}
