% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dataset}
\alias{plot_dataset}
\title{Generate  an IDOC plot}
\usage{
plot_dataset(
  experiment_folder,
  dataset,
  analysis,
  plot_preference_index = TRUE,
  plot_decision_zone = TRUE,
  subtitle = "",
  colors = c(TREATMENT_A = "red", TREATMENT_B = "blue"),
  ...
)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC hardware at one point in time
\item limits
\item border
}}

\item{analysis}{List of processed data including:
\itemize{
\item annotation: every row is a decision zone exit and its appetitive / aversive
value
\item pi: preference index scored by each animal
}}

\item{plot_preference_index}{Whether to show the scored preference index
with the region id on the facet label (TRUE), or just the region id (FALSE)}

\item{plot_decision_zone}{Whether to display the decision zone (TRUE) or not.}

\item{subtitle}{Character to write on the plot subtitles}

\item{colors}{Named vector of colors. Values should be colors
and names need to map to controller events}

\item{...}{Extra arguments for save_plot}
}
\description{
Take a dataset and its analysis,
together with plotting parameters,
to visualize an IDOC experiment
}
\seealso{
\code{\link[=mark_stimuli]{mark_stimuli()}}

\code{\link[=mark_decision_zone]{mark_decision_zone()}}

\code{\link[=mark_crosses]{mark_crosses()}}

\code{\link[=save_plot]{save_plot()}}
}
