% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference-index.R
\name{preference_index}
\alias{preference_index}
\title{Compute the preference index of a fly for the right side,
as defined by PAPER}
\usage{
preference_index(pos = NULL, min_exits_required = 5, min_length = 10)
}
\arguments{
\item{pos}{A character vector encoding the position of the fly in the chamber (L/D/R)}

\item{min_required}{Number of exits required to compute the index}
}
\value{
Numeric ranging 0-1 0 shows total aversion for odour on right side, and 1 total preference. -1 is return if min_required is not reached
}
\description{
Compute the preference index of a fly for the right side,
as defined by PAPER
}
\examples{
pos <- c('L','L','L','D','D','D','L','L',"L",'D','D','R','R','R','D')
preference_index(pos)
}
