% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_shape_data.R
\name{prepare_shape_data}
\alias{prepare_shape_data}
\title{Reshape the controller data to be plotted in ggplot2}
\usage{
prepare_shape_data(controller_data, hardware = "LED_R_LEFT")
}
\description{
Given a table containing when several hardware turned on/off
during an experiment, and a piece of hardware,
prepare_shape_data will generate a table with one row per corner in a rectangle
encompassing all the area of the plot where the hardware is active
This is given by its side (X axis) and the time when it was on (Y axis)
If the hardware was turned on/off n times, we get n rectangles and thus n*4 rows.#'
The x axis is assumed to go from -1 to 1. To scale it, see scale_shape
}
