% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-and-plot.R
\name{preprocess_and_plot}
\alias{preprocess_and_plot}
\title{Process the results of an experiment performed using the FSL learning memory setup}
\usage{
preprocess_and_plot(
  experiment_folder,
  index_function = LeMDTr::preference_index,
  decision_zone_mm = 10,
  min_exits_required = 5,
  max_time_minutes = Inf,
  annotation = "",
  A = NULL,
  B = NULL,
  selected_flies = 1:20,
  debug = FALSE
)
}
\arguments{
\item{experiment_folder}{character, path to results folder of an experiment}

\item{index_function}{function, returns an satistic based on a character vector of states}

\item{decision_zone_mm}{numeric, widht of the decision in mm}

\item{min_exits_required}{integer, minimum number of crossing events required to return an index}

\item{max_time_minutes}{numeric, drop all data after this amount of time since the start of the recording}

\item{annotation}{character, text to write in the top left corner of the plot}

\item{A}{character, name of odour in valves connected to bottle A}

\item{B}{character, name of odour in valves connected to bottle B}

\item{selected_flies}{integer vector, only these \code{region_id}s will be included in the plot}
}
\description{
Process the results of an experiment performed using the FSL learning memory setup
}
