% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{preprocess_controller}
\alias{preprocess_controller}
\title{Preprocess controller
\itemize{
\item Make sure the status at time 0 is present
\item Shift all the events to account for some delay in the arrival of the stimuli
if the user passes a delay
}}
\usage{
preprocess_controller(controller_data, delay = 0)
}
\arguments{
\item{controller_data}{data.frame with recordings of IDOC status at several timepoints,
one timepoint per row. Wide format}

\item{delay}{Seconds to shift the representation of the treatments
on the plot, to account for a latency or lag between treatment
delivery start and arrival to behavioral arena}
}
\description{
Preprocess controller
\itemize{
\item Make sure the status at time 0 is present
\item Shift all the events to account for some delay in the arrival of the stimuli
if the user passes a delay
}
}
