% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{preprocess_dataset}
\alias{preprocess_dataset}
\title{Preprocess a RAW idoc dataset}
\usage{
preprocess_dataset(
  experiment_folder,
  dataset,
  treatments = c("TREATMENT_A", "TREATMENT_B"),
  delay = 0,
  border_mm = 5,
  CSplus_idx = 1
)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}

\item{treatments}{Vector of treatments.
If named and names contain the word treatment,
it is likely that the user is running old API code}

\item{delay}{Seconds to shift the representation of the treatments
on the plot, to account for a latency or lag between treatment
delivery start and arrival to behavioral arena}

\item{border_mm}{mm from center of chamber to decision zone edge}

\item{CSplus_idx}{Whether the first treatment (CSplus_idx=1) or the second
(CSplus_idx=2) is associated with + exits}
}
\value{
tracker_data
}
\description{
\enumerate{
\item Preprocess tracker and controller data
\item Features expected by the analysis functions
are only available after some minor computations in the raw data
}
}
