% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{preprocess_tracker}
\alias{preprocess_tracker}
\title{Preprocess raw tracker data
\itemize{
\item Add a unique identifier to each animal
\item Center the space coordinates around 0
\item Keep only the needed columns
}}
\usage{
preprocess_tracker(experiment_folder, tracker_data)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{tracker_data}{IDOC tracker dataset where every row captures the position of one animal at one timepoint}
}
\value{
tracker_data
}
\description{
Preprocess raw tracker data
\itemize{
\item Add a unique identifier to each animal
\item Center the space coordinates around 0
\item Keep only the needed columns
}
}
