% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{process_folder}
\alias{process_folder}
\title{Document IDOC LTM experiment with ethoscope data}
\usage{
process_folder(folder)
}
\description{
Your IDOC folder must contain
\enumerate{
\item A valid metadata.cv file with metadata in the filename
\item An ID_reassignments.csv file with columns pre,post,etho
}
}
