% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_load.R
\name{read_pi}
\alias{read_pi}
\title{Read the Preference Index (PI) achieved by a fly in a trial or set of trials
Load the PI as computed by idocr (machine) and stored in a .csv file
The PI will be set to NA if only aversive or appetitive exits occur,
AND also if their sum is not at least min_exits
n_exits will be NA if only aversive or appetitive exits occur, otherwise it will be their sum
(even if under min_exits)}
\usage{
read_pi(path, roi, min_exits = 3)
}
\description{
Read the Preference Index (PI) achieved by a fly in a trial or set of trials
Load the PI as computed by idocr (machine) and stored in a .csv file
The PI will be set to NA if only aversive or appetitive exits occur,
AND also if their sum is not at least min_exits
n_exits will be NA if only aversive or appetitive exits occur, otherwise it will be their sum
(even if under min_exits)
}
