% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{rebound}
\alias{rebound}
\title{Generate a toy dataset for testing purposes
Rebound movement if the animal hits the minimum of maximum of the dimension}
\usage{
rebound(x, minn, maxx)
}
\arguments{
\item{x}{Current position on either dimension (length 1)}

\item{minn}{Minimum value on that dimension}

\item{maxx}{Maximum value on that dimension}
}
\value{
Corrected position with min/max offset rebounded into allowed bounds
}
\description{
Useful so movement magnitude is not lost due to wall efect
}
