% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate entries in a data table
Duplicates have same region_id and t}
\usage{
remove_duplicates(tracker_data)
}
\arguments{
\item{tracker_data}{IDOC tracker dataset where every row captures the position of one animal at one timepoint}
}
\description{
Remove duplicate entries in a data table
Duplicates have same region_id and t
}
