% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{remove_empty_roi}
\alias{remove_empty_roi}
\title{Undo the effect of add_empty_roi}
\usage{
remove_empty_roi(tracker_data, minimum = 100)
}
\arguments{
\item{tracker_data}{IDOC tracker dataset where every row captures the position of one animal at one timepoint}

\item{minimum}{Minimum number of datapoints in a row to be considered
not sparse (not noise) and thus valid}
}
\description{
Useful to remove dummy rois or rois with noise
Works only if not testing. During testing
we may want to pass small (noise-like) datasets
}
