% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot2 plot in pdf and png format using consistent naming}
\usage{
save_plot(gg, experiment_folder, result_folder = NULL, suffix = "", ...)
}
\arguments{
\item{gg}{ggplot2 plot}

\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{result_folder}{Path to a folder where to save results. If not passaed, it is set to experiment_folder}

\item{suffix}{A character (word) that will differentiate resulting output
from other output produced on the same experiment folder.
Useful for segmenting the experiment into different periods of interest
and analyzing these periods separately.}

\item{...}{Extra arguments for ggsave}
}
\description{
Save a ggplot2 plot in pdf and png format using consistent naming
}
\seealso{
\code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
}
