% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{scale_rectangle}
\alias{scale_rectangle}
\title{Scale the x axis of a rectangle data table
using the x column of tracking data
scale_shape fetches the width of the chambers
and scales the shape accodingly}
\usage{
scale_rectangle(shape_data, limits, border)
}
\description{
Scale the x axis of a rectangle data table
using the x column of tracking data
scale_shape fetches the width of the chambers
and scales the shape accodingly
}
