% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scale_shape}
\alias{scale_shape}
\title{Scale the x axis of a shape data table
using the x column of tracking data
scale_shape fetches the width of the chambers
and scales the shape accodingly}
\usage{
scale_shape(shape_data, limits, border)
}
\description{
Scale the x axis of a shape data table
using the x column of tracking data
scale_shape fetches the width of the chambers
and scales the shape accodingly
}
