% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{seconds_mask}
\alias{seconds_mask}
\title{Dont count crosses happening within less than seconds_masked seconds
since the previous one}
\usage{
seconds_mask(data, seconds_masked = 5)
}
\description{
Dont count crosses happening within less than seconds_masked seconds
since the previous one
}
