% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sort_facet_levels}
\alias{sort_facet_levels}
\title{Enforce a numeric sorting of facets for visualization}
\usage{
sort_facet_levels(facets)
}
\arguments{
\item{facets}{Character of facets to be sorted}
}
\value{
Facet levels sorted in numeric order
(ROI_1, ROI_2, ..., ROI_19, ROI_20)
}
\description{
Enforce a numeric sorting of facets for visualization
}
