% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{toy_roi}
\alias{toy_roi}
\title{Generate a toy fly movement dataset}
\usage{
toy_roi(steps = 100, p = 0.2, ...)
}
\arguments{
\item{steps}{Number of timepoints in toy dataset}

\item{p}{Probability of a quiescent fly i.e. p chance that the simulated
fly does not move at all}

\item{...}{extra arguments to walk}
}
\description{
Generate a toy fly movement dataset
}
