% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{toy_roi_all}
\alias{toy_roi_all}
\title{Generate a multi animal dataset}
\usage{
toy_roi_all(channels = 20, ...)
}
\arguments{
\item{channels}{Number of animals to simulate}

\item{...}{Extra arguments to \code{\link[=toy_roi]{toy_roi()}}}
}
\description{
Call toy_roi as many times as channels are passed
}
\seealso{
\code{\link[=toy_roi]{toy_roi()}}
}
