% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{update_scores_with_manual_annotation}
\alias{update_scores_with_manual_annotation}
\title{Manually override scores with a human-made score
The manual score is provided in the PRE and POST columns of the metadata
and applied only if PRE_Reason and POST_reason have a reason other than empty, ? and Machine-override}
\usage{
update_scores_with_manual_annotation(data, scoring_system)
}
\description{
Manually override scores with a human-made score
The manual score is provided in the PRE and POST columns of the metadata
and applied only if PRE_Reason and POST_reason have a reason other than empty, ? and Machine-override
}
