% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{validate_events}
\alias{validate_events}
\title{Given a dataset, find the timepoints
where the stimulus changes state}
\usage{
validate_events(data, stimulus)
}
\description{
Given a dataset, find the timepoints
where the stimulus changes state
}
