% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_number_of_fields}
\alias{validate_number_of_fields}
\title{Make sure the passed csv file has same number of fields on every row}
\usage{
validate_number_of_fields(csv_file)
}
\arguments{
\item{csv_file}{Path to a .csv file produced by IDOC}
}
\description{
Make sure the passed csv file has same number of fields on every row
}
