% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{validate_varmap_file}
\alias{validate_varmap_file}
\title{Confirm VAR_MAP csv table is not corrupted}
\usage{
validate_varmap_file(var_map_path)
}
\arguments{
\item{var_map_path}{Path to a VAR_MAP table .csv file}
}
\description{
Confirm VAR_MAP csv table is not corrupted
}
