% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{walk}
\alias{walk}
\title{Brownian motion on the 2d space}
\usage{
walk(start_pos, quiescent = FALSE, sd_y = 0.5)
}
\arguments{
\item{start_pos}{numeric vector of length 2 with current position of animal in x and y coords}

\item{quiescent}{Whether the fly can walk (FALSE) or not (TRUE)}

\item{sd_y}{standard deviation of the normal distribution governing the movement on y}
}
\description{
Brownian motion on the 2d space
}
