% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{write_dataset}
\alias{write_dataset}
\title{Save a dataset to a .csv database
A .csv database is a collection of .csv files under the same folder
where every .csv file has a common prefix and a differential key
that distinguishes it from other files and hints at its contents}
\usage{
write_dataset(dataset, dest)
}
\arguments{
\item{dataset}{A list with entries metadata, roi_data,
controller_data, var_map and roi_map}

\item{dest}{A folder where to generate a .csv database.
Created on the spot if it doesn't exist}
}
\description{
Save a dataset to a .csv database
A .csv database is a collection of .csv files under the same folder
where every .csv file has a common prefix and a differential key
that distinguishes it from other files and hints at its contents
}
