% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{add_empty_roi}
\alias{add_empty_roi}
\title{Represent missing ROIs in the dataset with a single row}
\usage{
add_empty_roi(experiment_folder, tracker_data, n = 20)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{tracker_data}{IDOC tracker dataset where every row captures the position of one animal at one timepoint}

\item{n}{Number of ROIS the output must have (empty or not)}
}
\description{
This is necessary so the results are always for 20 channels
regardless of whether the channel had an animal or not.
Doing so makes the results systematic and easily comparable across runs
}
