% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_empty_roi.R
\name{add_empty_roi}
\alias{add_empty_roi}
\title{Represent missing ROIs in the dataset with a single row}
\usage{
add_empty_roi(experiment_folder, roi_data, n = 20)
}
\description{
This is done to show an empty facet that however fits
the expected n animals
}
