% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse_dataset}
\alias{analyse_dataset}
\title{Analyse an IDOC dataset}
\usage{
analyse_dataset(dataset, min_exits_required = 5, ...)
}
\arguments{
\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}

\item{min_exits_required}{Minimal number of exits
to be considered for significance}

\item{...}{Extra arguments to find_exits_all}
}
\description{
Take an IDOC dataset
}
\seealso{
\code{\link[=annotate_cross]{annotate_cross()}}
}
