% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{annotate_cross}
\alias{annotate_cross}
\title{Annotate registered crosses with either appetitive or aversive behavior}
\usage{
annotate_cross(
  cross_data,
  event_data,
  treatment,
  type = c("appetitive", "aversive")
)
}
\arguments{
\item{cross_data}{data.frame where each row is an exit, defined by region_id, t and side}

\item{event_data}{data.frame where each row is an event, defined by stimulus, t_start t_end, side}

\item{treatment}{Name of a treatment matching a treatment in the corresponding field of \code{event_data}}

\item{type}{type of event. Either appetitive, which makes the PI increase
when the fly exits the decision zone towards the event,
and aversive, if the PI should decrease in that case}
}
\value{
data.frame of crosses happening during some event and with annotated type
}
\description{
Annotate registered crosses with either appetitive or aversive behavior
}
