% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{annotate_cross_all}
\alias{annotate_cross_all}
\title{Wrapper around annotate_cross for appetitive and aversive conditioning}
\usage{
annotate_cross_all(cross_data, event_data, CSplus, CSminus)
}
\description{
Wrapper around annotate_cross for appetitive and aversive conditioning
}
