% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{base_plot}
\alias{base_plot}
\title{Generate base IDOC plot upon which
more annotation layers can be added}
\usage{
base_plot(
  gg,
  data,
  limits,
  line_alpha = 1,
  downward = TRUE,
  nrow = 1,
  ncol = 20,
  orientation = "y"
)
}
\arguments{
\item{gg}{ggplot2 object}

\item{data}{}

\item{limits}{Position of ticks in space axis}

\item{line_alpha}{Alpha of position trace}

\item{downward}{Whether the 0 should be at the top (TRUE)
or the bottom (FALSE) of the plot}

\item{nrow}{Number of rows used for facetting data}

\item{ncol}{Number of cols used for facetting data}
}
\value{
gg ggplot2 object
}
\description{
The base plot contains all individual panels,
dots to mark the fly positions
}
\seealso{
\code{\link[=mark_space]{mark_space()}}

\code{\link[=mark_time]{mark_time()}}
}
