% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{base_plot}
\alias{base_plot}
\title{Generate base IDOC plot upon which
more annotation layers can be added}
\usage{
base_plot(
  experiment_folder,
  data,
  limits,
  pi,
  run_id = NULL,
  plot_preference_index = TRUE,
  subtitle = ""
)
}
\description{
Generate base IDOC plot upon which
more annotation layers can be added
}
