% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{build_filename}
\alias{build_filename}
\title{Moduke to infer names of .csv files based on metadata+key or experiment_folder+key
#' Build a filename based on the metadata and a key}
\usage{
build_filename(metadata, key)
}
\arguments{
\item{metadata}{data.table with all information about the experiment
i.e. when it was done, etc}

\item{key}{Name of a table in the resulting database}
}
\value{
Filename with systematic metainformation and key to differentiate
it from other files from the same experiment
}
\description{
Moduke to infer names of .csv files based on metadata+key or experiment_folder+key
#' Build a filename based on the metadata and a key
}
