% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{combine_inputs}
\alias{combine_inputs}
\title{Distill the dataset and analysis performed in the idocr workflow
to the datasets required for plotting}
\usage{
combine_inputs(dataset, analysis, plot_preference_index = TRUE)
}
\description{
A tracking dataset where every record represents the position
of one animal in one timepoint is required to plot the trace
}
\details{
A corssing dataset where every record represents a decision zone
exit of one animal is required to show the exits in the plot
}
