% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{define_rectangles}
\alias{define_rectangles}
\title{Wrapper around define_rectangle for several hardware}
\usage{
define_rectangles(controller_data, hardware)
}
\arguments{
\item{controller_data}{A wide data.frame of hardware and a t column}

\item{hardware}{A character vector consisting of a subset of the columns in controller_data
for which rectangles describing the period when the hardware was on, should be built}
}
\description{
Wrapper around define_rectangle for several hardware
}
