% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_summary}
\alias{export_summary}
\title{Export a single csv with all key data for analysis and plotting outside of R}
\usage{
export_summary(experiment_folder, output_csv = NULL, ...)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{output_csv}{Path of csv output. If not provided,}

\item{...}{Extra arguments to make_summary}
}
\description{
Export a single csv with all key data for analysis and plotting outside of R
}
\seealso{
\code{\link[=make_summary]{make_summary()}}
a new file in the experiment folder with key SUMMARY is produced
}
