% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_summary.R
\name{export_summary}
\alias{export_summary}
\title{Export a single csv with all key data for analysis and plotting outside of R}
\usage{
export_summary(experiment_folder, output_csv)
}
\description{
Export a single csv with all key data for analysis and plotting outside of R
}
