% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_data.R
\name{get_event_data}
\alias{get_event_data}
\title{Reshape controller data to a format where it is
easy to merge with cross_data}
\usage{
get_event_data(controller_data)
}
\description{
Reshape controller data to a format where it is
easy to merge with cross_data
}
