% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idocr.R
\name{idocr}
\alias{idocr}
\title{The standard idocr workflow.}
\usage{
idocr(
  experiment_folder,
  hardware = c("LED_R_LEFT", "LED_R_RIGHT"),
  old_mapping = FALSE,
  plot_basename = NULL,
  border_mm = 5,
  min_exits_required = 5
)
}
\description{
Load data from the experiment folder, possibly with some specific settings
and return a data frame with the preference index of each fly
and a plot visualizing the experiment
}
