% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{load_controller}
\alias{load_controller}
\title{Load a CONTROLLER_EVENTS .csv table into R}
\usage{
load_controller(experiment_folder, ...)
}
\arguments{
\item{experiment_folder}{Path to a folder with IDOC results}

\item{delay}{Shift this amount of seconds the onset of the stimuli forward in time
to account for delays in the arrival of the stimuli (stimuli delivery is not immediate)}
}
\description{
Load a CONTROLLER_EVENTS .csv table into R
}
