% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_dataset}
\alias{load_dataset}
\title{Load an IDOC .csv dataset}
\usage{
load_dataset(experiment_folder, ...)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{...}{Arguments to preprocess_dataset}
}
\description{
Load the .csv database available at the passed directory
}
\seealso{
\code{\link[=preprocess_dataset]{preprocess_dataset()}}
}
