% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{load_systematic_rois}
\alias{load_systematic_rois}
\title{Load data from n channels
and expose missing data if needed}
\usage{
load_systematic_rois(..., n = 20)
}
\arguments{
\item{...}{Arguments to load_rois}

\item{n}{Number of channels desired.
If less than this amount of animals is found, idocr
will make up animals until this amount is hit
It (should match nrow x ncol of layout passed in the idocr function)}
}
\value{
tibble
}
\description{
Load data from n channels
and expose missing data if needed
}
