% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_crosses}
\alias{mark_crosses}
\title{Mark decision zone exit events (crosses)}
\usage{
mark_crosses(gg, data, analysis)
}
\arguments{
\item{gg}{ggplot2 object}

\item{data}{List with entries roi, controller, border, limits.
roi and controller must be data.tables generated by loading IDOC data.
Border is a TODO. Limits is a numeric vector with 2 positions, delineating
the boundaries of the x axis}

\item{analysis}{List of processed data including:
\itemize{
\item annotation: every row is a decision zone exit and its appetitive / aversive
value
\item pi: preference index scored by each animal
}}
}
\description{
Mark decision zone exit events (crosses)
}
