% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_crosses}
\alias{mark_crosses}
\title{Mark decision zone exit events (crosses)}
\usage{
mark_crosses(
  gg,
  cross_data,
  size = 2,
  color = "black",
  orientation = "y",
  style = "default"
)
}
\arguments{
\item{gg}{ggplot2 object}

\item{cross_data}{data.frame where every row captures the position of one animal at one timepoint
when the animal exits the decision zone. Must contain fields t, side, region_id}

\item{size}{Size of markers representing crosses}

\item{color}{Color of the points, black by default}
}
\description{
Mark decision zone exit events (crosses)
}
