% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mark_stimuli}
\alias{mark_stimuli}
\title{Mark stimuli / treatment action over time with rectangles on the plot}
\usage{
mark_stimuli(gg, rectangles, colors, labels)
}
\arguments{
\item{gg}{ggplot2 object}

\item{rectangles}{List of rectangle datasets where every dataset is a
dataframe of 4 rows describing the 4 corners of a rectangle.
They represent the occurrence of a stimulus each#'}

\item{colors}{Named character vector where values are colors and names
are treatments. It establishes the color used to represent the treatments
on the plot}

\item{labels}{Character vector whose values become
the name of the treatments as rendered in the plot's legend}
}
\value{
ggplot2 object
}
\description{
Mark stimuli / treatment action over time with rectangles on the plot
}
