% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dataset}
\alias{plot_dataset}
\title{Generate  an IDOC plot}
\usage{
plot_dataset(
  experiment_folder,
  dataset,
  analysis,
  result_folder = NULL,
  plot_preference_index = TRUE,
  plot_decision_zone = TRUE,
  plot_crosses = TRUE,
  subtitle = "",
  colors = c(TREATMENT_A = "red", TREATMENT_B = "blue"),
  labels = c(TREATMENT_A = "TREATMENT_A", TREATMENT_B = "TREATMENT_B"),
  analysis_mask = NULL,
  plot_mask = NULL,
  downward = TRUE,
  nrow = 1,
  ncol = 20,
  suffix = "",
  cross_size = 2,
  line_alpha = 1,
  ...
)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}

\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}

\item{analysis}{List of processed data including:
\itemize{
\item annotation: every row is a decision zone exit and its appetitive / aversive
value
\item pi: preference index scored by each animal
}}

\item{result_folder}{Path to a folder where to save results. If not passaed, it is set to experiment_folder}

\item{plot_preference_index}{Whether to show the scored preference index
with the region id on the facet label (TRUE), or just the region id (FALSE)}

\item{plot_decision_zone}{Whether to display the decision zone (TRUE) or not.}

\item{plot_crosses}{Whether to display the decision zone crosses (TRUE) or not.}

\item{subtitle}{Character to write on the plot subtitles}

\item{colors}{Named vector of colors. Values should be colors
and names need to map to controller events}

\item{labels}{Character vector whose values become
the name of the treatments as rendered in the plot's legend}

\item{analysis_mask}{Numeric vector of length 2
containing start and end timepoint of mask
in seconds since the start of the experiment
Every behavior outside of the mask is ignored}

\item{plot_mask}{If not NULL, the plot contains data contained within
the interval's start and end (s)}

\item{downward}{Whether the 0 should be at the top (TRUE)
or the bottom (FALSE) of the plot}

\item{nrow}{Number of rows used for facetting data}

\item{ncol}{Number of cols used for facetting data}

\item{suffix}{A character (word) that will differentiate resulting output
from other output produced on the same experiment folder.
Useful for segmenting the experiment into different periods of interest
and analyzing these periods separately.}

\item{cross_size}{Size of dots used to represent decision zone exits (crosses)}

\item{line_alpha}{Alpha of position trace}

\item{...}{Extra arguments for save_plot}
}
\value{
ggplot2 object
}
\description{
Take a dataset and its analysis,
together with plotting parameters,
to visualize an IDOC experiment
}
\details{
A corssing dataset where every record represents a decision zone
exit of one animal is required to show the exits in the plot
}
\seealso{
\code{\link[=mark_stimuli]{mark_stimuli()}}

\code{\link[=mark_decision_zone]{mark_decision_zone()}}

\code{\link[=mark_crosses]{mark_crosses()}}

\code{\link[=save_plot]{save_plot()}}
}
