% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference.R
\name{preference_index}
\alias{preference_index}
\title{Compute the preference index of a dataset}
\usage{
preference_index(appetitive, aversive, min_exits_required = 5)
}
\arguments{
\item{appetitive}{Number of exits towards the CSplus simulus}

\item{aversive}{Number of exits towards the CSminus stimulus}

\item{min_exits_required}{An integer stating the minimal number
of cross events required for the data to be considered significant}
}
\description{
The preference index is a measurement of how much
an animal preferes to be exposed to one stimulus
over another differnt stimulus or nothing.
If stimuli are applied on extremes of the chamber,
the position of the fly along the x axis of the chamber
can be used as proxy for preference.
The preference index is built upon the number of crosses
of a so called "decision zone" situated in the center of the chamber
Leaving the decision zone it in the direction of the stimulus
are considered evidence of preference for it, whereas
Leaving the decision zone on the opposite direction
has the opposite effect. The preference index takes the difference
between both and normalizes for the total number of crosses.
A minimum number of crosses can be required for the whole dataset
or for each exposure to the stimulus.
}
