% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{preprocess_controller}
\alias{preprocess_controller}
\title{Preprocess controller
\itemize{
\item Make sure the status at time 0 is present
\item Shift all the events to account for some delay in the arrival of the stimuli
if the user passes a delay
}}
\usage{
preprocess_controller(controller_data, delay = 0)
}
\description{
Preprocess controller
\itemize{
\item Make sure the status at time 0 is present
\item Shift all the events to account for some delay in the arrival of the stimuli
if the user passes a delay
}
}
