% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{preprocess_dataset}
\alias{preprocess_dataset}
\title{Preprocess a RAW idoc dataset}
\usage{
preprocess_dataset(
  experiment_folder,
  dataset,
  treatments = c("TREATMENT_A", "TREATMENT_B"),
  delay = 0,
  border_mm = 5,
  CSplus_idx = 1
)
}
\arguments{
\item{experiment_folder}{Path to folder containing IDOC results .csv files}
}
\value{
tracker_data
}
\description{
\enumerate{
\item Preprocess tracker and controller data
\item Features expected by the analysis functions
are only available after some minor computations in the raw data
}
}
