% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{reshape_controller}
\alias{reshape_controller}
\title{Reshape controller data to a format where it is
easy to merge with cross_data}
\usage{
reshape_controller(rectangle_data)
}
\description{
Enter a dataframe with one row per corner
get a new dataframe with one row per event (1/4 of the rows)
}
