% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot2 plot in pdf and png format using consistent naming}
\usage{
save_plot(gg, experiment_folder, ...)
}
\arguments{
\item{gg}{ggplot2 plot}

\item{experiment_folder}{Path to folder containing IDOC results .csv files}
}
\description{
Save a ggplot2 plot in pdf and png format using consistent naming
}
