% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{seconds_mask}
\alias{seconds_mask}
\title{Ignore too close crosses}
\usage{
seconds_mask(cross_data, min_time = 0)
}
\arguments{
\item{cross_data}{data.frame of crosses}

\item{min_time}{Minimum amount of seconds between crosses allowed}
}
\description{
Dont count crosses happening within less than seconds_masked seconds
since the previous one
}
