% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{toy_controller}
\alias{toy_controller}
\title{Generate a toy controller dataset}
\usage{
toy_controller(paradigm = NULL)
}
\arguments{
\item{paradigm}{data.table with format hardware, start, off}
}
\value{
data.table with one column per hardware in paradigm
and one row per sampling timepoint. The value of the i,j cell
states whether the ith hardware was on at time j
}
\description{
Generate a toy controller dataset
}
