% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangles.R
\name{validate_events}
\alias{validate_events}
\title{Given a dataset, find the timepoints
where the stimulus changes state}
\usage{
validate_events(controller_data, stimulus)
}
\arguments{
\item{controller_data}{data.frame with recordings of IDOC status at several timepoints,
one timepoint per row. Wide format}

\item{stimulus}{Namew of stimulus or event, which should match
a column in the passed \code{controller_data}}
}
\description{
Given a dataset, find the timepoints
where the stimulus changes state
}
