% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate data passed to plot_dataset}
\usage{
validate_inputs(dataset, analysis)
}
\arguments{
\item{dataset}{List of raw data including:
\itemize{
\item roi: every row records the position of one animal at one point in time
\item controller: every row records the status of all IDOC stimulus at one point in time
\item limits
\item border
}}

\item{analysis}{List of processed data including:
\itemize{
\item annotation: every row is a decision zone exit and its appetitive / aversive
value
\item pi: preference index scored by each animal
}}
}
\value{
ggplot2 object
}
\description{
Take a dataset and its analysis,
together with plotting parameters,
to visualize an IDOC experiment
}
\seealso{
\code{\link[=mark_stimuli]{mark_stimuli()}}

\code{\link[=mark_decision_zone]{mark_decision_zone()}}

\code{\link[=mark_crosses]{mark_crosses()}}

\code{\link[=save_plot]{save_plot()}}
}
